Divided Linux and Windows tracing logic into separate classes
